requirements.txt
# Trading Bot Requirements
# Core dependencies for algorithmic trading bot

# Market data and exchange integration
ccxt>=4.0.0              # Cryptocurrency exchange library
yfinance>=0.2.0          # Yahoo Finance data (S&P 500, stocks)
pandas>=2.0.0            # Data manipulation
numpy>=1.24.0            # Numerical computing

# S&P 500 and Traditional Markets
alpaca-trade-api>=3.0.0  # Alpaca broker API (commission-free stocks)
# ib-insync>=0.9.86       # Interactive Brokers API (uncomment if using IBKR)
# schwab-api>=1.0.0       # Charles Schwab API (uncomment if using Schwab)

# Technical analysis
TA-Lib>=0.4.25           # Technical analysis library
# pandas-ta>=0.3.14b       # Pandas technical analysis (optional)

# Web framework and APIs
fastapi>=0.104.0         # REST API framework
uvicorn>=0.24.0          # ASGI server
websockets>=12.0         # WebSocket support
aiohttp>=3.9.0           # Async HTTP client

# Database and storage
sqlalchemy>=2.0.0        # Database ORM
alembic>=1.12.0          # Database migrations
redis>=5.0.0             # Caching and message queue
psycopg2-binary>=2.9.0   # PostgreSQL adapter

# Data science and ML
scikit-learn>=1.3.0      # Machine learning
scipy>=1.11.0            # Scientific computing
matplotlib>=3.7.0        # Plotting
seaborn>=0.12.0          # Statistical visualization
plotly>=5.17.0           # Interactive plotting

# Configuration and utilities
pydantic>=2.5.0          # Data validation
pyyaml>=6.0              # YAML configuration
python-dotenv>=1.0.0     # Environment variables
schedule>=1.2.0          # Task scheduling

# Logging and monitoring
loguru>=0.7.0            # Advanced logging
prometheus-client>=0.19.0 # Metrics collection

# Testing
pytest>=7.4.0           # Testing framework
pytest-asyncio>=0.21.0  # Async testing
pytest-cov>=4.1.0       # Coverage reporting

# Jupyter and analysis
jupyter>=1.0.0          # Notebook interface
ipywidgets>=8.1.0       # Interactive widgets

# Security and encryption
cryptography>=41.0.0    # Encryption utilities
keyring>=24.3.0         # Secure credential storage

# Development tools
black>=23.0.0           # Code formatting  
flake8>=6.1.0           # Code linting
mypy>=1.7.0             # Type checking